refactor (divide module)
