+++ src/main/java/racing/Strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.Strategy;
 
 public interface MoveStrategy {
