+++ src/main/java/racing/View/ResultView.java
@@ -1,4 +1,6 @@
-package racing;
+package racing.View;
+
+import racing.game.Car;
 
 import java.util.List;
