+++ src/main/java/racing/game/Car.java
@@ -1,8 +1,9 @@
-package racing;
+package racing.game;
+
+import racing.Strategy.MoveStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
 
 public class Car {
