+++ src/test/java/racing/game/CarTest.java
@@ -1,7 +1,9 @@
-package racing;
+package racing.game;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.Strategy.MoveStrategy;
+import racing.game.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
